Separate gathering logic
