+++ pkg/gather/clusterconfig/authentications.go
@@ -1,0 +1,30 @@
+package clusterconfig
+
+import (
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	_ "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
+
+	"github.com/openshift/insights-operator/pkg/record"
+)
+
+// GatherClusterAuthentication fetches the cluster Authentication - the Authentication with name cluster.
+//
+// The Kubernetes api https://github.com/openshift/client-go/blob/master/config/clientset/versioned/typed/config/v1/authentication.go#L50
+// Response see https://docs.openshift.com/container-platform/4.3/rest_api/index.html#authentication-v1operator-openshift-io
+//
+// Location in archive: config/authentication/
+// See: docs/insights-archive-sample/config/authentication
+func GatherClusterAuthentication(i *Gatherer) func() ([]record.Record, []error) {
+	return func() ([]record.Record, []error) {
+		config, err := i.client.Authentications().Get(i.ctx, "cluster", metav1.GetOptions{})
+		if errors.IsNotFound(err) {
+			return nil, nil
+		}
+		if err != nil {
+			return nil, []error{err}
+		}
+		return []record.Record{{Name: "config/authentication", Item: Anonymizer{config}}}, nil
+	}
+}
