+++ pkg/gather/clusterconfig/feature_gates.go
@@ -1,0 +1,47 @@
+package clusterconfig
+
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+
+	configv1 "github.com/openshift/api/config/v1"
+	_ "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
+
+	"github.com/openshift/insights-operator/pkg/record"
+)
+
+// GatherClusterFeatureGates fetches the cluster FeatureGate - the FeatureGate with name cluster.
+//
+// The Kubernetes api https://github.com/openshift/client-go/blob/master/config/clientset/versioned/typed/config/v1/featuregate.go#L50
+// Response see https://docs.openshift.com/container-platform/4.3/rest_api/index.html#featuregate-v1-config-openshift-io
+//
+// Location in archive: config/featuregate/
+// See: docs/insights-archive-sample/config/featuregate
+func GatherClusterFeatureGates(i *Gatherer) func() ([]record.Record, []error) {
+	return func() ([]record.Record, []error) {
+		config, err := i.client.FeatureGates().Get(i.ctx, "cluster", metav1.GetOptions{})
+		if errors.IsNotFound(err) {
+			return nil, nil
+		}
+		if err != nil {
+			return nil, []error{err}
+		}
+		return []record.Record{{Name: "config/featuregate", Item: FeatureGateAnonymizer{config}}}, nil
+	}
+}
+
+// FeatureGateAnonymizer implements serializaton of FeatureGate with anonymization
+type FeatureGateAnonymizer struct{ *configv1.FeatureGate }
+
+// Marshal serializes FeatureGate with anonymization
+func (a FeatureGateAnonymizer) Marshal(_ context.Context) ([]byte, error) {
+	return runtime.Encode(openshiftSerializer, a.FeatureGate)
+}
+
+// GetExtension returns extension for anonymized feature gate objects
+func (a FeatureGateAnonymizer) GetExtension() string {
+	return "json"
+}
