+++ pkg/gather/clusterconfig/oauths.go
@@ -1,0 +1,30 @@
+package clusterconfig
+
+import (
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	_ "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
+
+	"github.com/openshift/insights-operator/pkg/record"
+)
+
+// GatherClusterOAuth fetches the cluster OAuth - the OAuth with name cluster.
+//
+// The Kubernetes api https://github.com/openshift/client-go/blob/master/config/clientset/versioned/typed/config/v1/oauth.go#L50
+// Response see https://docs.openshift.com/container-platform/4.3/rest_api/index.html#oauth-v1-config-openshift-io
+//
+// Location in archive: config/oauth/
+// See: docs/insights-archive-sample/config/oauth
+func GatherClusterOAuth(i *Gatherer) func() ([]record.Record, []error) {
+	return func() ([]record.Record, []error) {
+		config, err := i.client.OAuths().Get(i.ctx, "cluster", metav1.GetOptions{})
+		if errors.IsNotFound(err) {
+			return nil, nil
+		}
+		if err != nil {
+			return nil, []error{err}
+		}
+		return []record.Record{{Name: "config/oauth", Item: Anonymizer{config}}}, nil
+	}
+}
