+++ pkg/gather/clusterconfig/proxies.go
@@ -1,0 +1,54 @@
+package clusterconfig
+
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+
+	configv1 "github.com/openshift/api/config/v1"
+	_ "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
+
+	"github.com/openshift/insights-operator/pkg/record"
+)
+
+// GatherClusterProxy fetches the cluster Proxy - the Proxy with name cluster.
+//
+// The Kubernetes api https://github.com/openshift/client-go/blob/master/config/clientset/versioned/typed/config/v1/proxy.go#L30
+// Response see https://docs.openshift.com/container-platform/4.3/rest_api/index.html#proxy-v1-config-openshift-io
+//
+// Location in archive: config/proxy/
+// See: docs/insights-archive-sample/config/proxy
+func GatherClusterProxy(i *Gatherer) func() ([]record.Record, []error) {
+	return func() ([]record.Record, []error) {
+		config, err := i.client.Proxies().Get(i.ctx, "cluster", metav1.GetOptions{})
+		if errors.IsNotFound(err) {
+			return nil, nil
+		}
+		if err != nil {
+			return nil, []error{err}
+		}
+		return []record.Record{{Name: "config/proxy", Item: ProxyAnonymizer{config}}}, nil
+	}
+}
+
+// ProxyAnonymizer implements serialization of HttpProxy/NoProxy with anonymization
+type ProxyAnonymizer struct{ *configv1.Proxy }
+
+// Marshal implements Proxy serialization with anonymization
+func (a ProxyAnonymizer) Marshal(_ context.Context) ([]byte, error) {
+	a.Proxy.Spec.HTTPProxy = anonymizeURLCSV(a.Proxy.Spec.HTTPProxy)
+	a.Proxy.Spec.HTTPSProxy = anonymizeURLCSV(a.Proxy.Spec.HTTPSProxy)
+	a.Proxy.Spec.NoProxy = anonymizeURLCSV(a.Proxy.Spec.NoProxy)
+	a.Proxy.Spec.ReadinessEndpoints = anonymizeURLSlice(a.Proxy.Spec.ReadinessEndpoints)
+	a.Proxy.Status.HTTPProxy = anonymizeURLCSV(a.Proxy.Status.HTTPProxy)
+	a.Proxy.Status.HTTPSProxy = anonymizeURLCSV(a.Proxy.Status.HTTPSProxy)
+	a.Proxy.Status.NoProxy = anonymizeURLCSV(a.Proxy.Status.NoProxy)
+	return runtime.Encode(openshiftSerializer, a.Proxy)
+}
+
+// GetExtension returns extension for anonymized proxy objects
+func (a ProxyAnonymizer) GetExtension() string {
+	return "json"
+}
